Feb 7, 2025, 10:25 AM

+++ CT06_04/lesson4.py
@@ -11,7 +11,7 @@ print("Noor Daanya is paying $" + str(customer1) + " to Mr Hong Guozhong for thi
 print("Khoo Siew Ling is paying $" + str(customer2) + " to Mr Huang Jonkin for this.")
 print("Smahon Loh is paying $" + str(customer3) + " to Ms Celeste Loh for this.")
 
-Smahon_Loh_money = int(input("Mrs Smahon, how much do you have in your wallet?"))
+Smahon_Loh_money = float(input("Mrs Smahon, how much do you have in your wallet?"))
 Smahon_Loh_money_1_hour_later = Smahon_Loh_money - customer3
 Smahon_Loh_money_1_year_later = Smahon_Loh_money_1_hour_later * 1.0425
 print("You will have $" + str(Smahon_Loh_money_1_year_later) + " next year.")
